wowai/requirements.txt
djangorestframework
django==4.1
mysqlclient